MATH-64: refactored abstract number class
