+++ orionmath_algebra/orionmath_algebra_number/src/test/java/io/github/orionlibs/math/algebra/number/NumTest_value.java
@@ -20,4 +20,34 @@ public class NumTest_value
         assertTrue(Num.of("0").isZero());
         assertTrue(Num.of().isZero());
     }
+
+
+    @Test
+    void test_hasPositiveInfiniteValue()
+    {
+        assertFalse(Num.of("999999999999999999999999999999999999999999999999999999999999999999999999999999.0").hasPositiveInfiniteValue());
+        assertTrue(Num.of("999999999999999999999999999999999999999999999999999999999999999999999999999999.1").hasPositiveInfiniteValue());
+        assertTrue(Num.of("9999999999999999999999999999999999999999999999999999999999999999999999999999999.0").hasPositiveInfiniteValue());
+    }
+
+
+    @Test
+    void test_hasNegativeInfiniteValue()
+    {
+        assertFalse(Num.of("-999999999999999999999999999999999999999999999999999999999999999999999999999999.0").hasNegativeInfiniteValue());
+        assertTrue(Num.of("-999999999999999999999999999999999999999999999999999999999999999999999999999999.1").hasNegativeInfiniteValue());
+        assertTrue(Num.of("-9999999999999999999999999999999999999999999999999999999999999999999999999999999.0").hasNegativeInfiniteValue());
+    }
+
+
+    @Test
+    void test_hasInfiniteValue()
+    {
+        assertFalse(Num.of("-999999999999999999999999999999999999999999999999999999999999999999999999999999.0").hasInfiniteValue());
+        assertTrue(Num.of("-999999999999999999999999999999999999999999999999999999999999999999999999999999.1").hasInfiniteValue());
+        assertTrue(Num.of("-9999999999999999999999999999999999999999999999999999999999999999999999999999999.0").hasInfiniteValue());
+        assertFalse(Num.of("999999999999999999999999999999999999999999999999999999999999999999999999999999.0").hasInfiniteValue());
+        assertTrue(Num.of("999999999999999999999999999999999999999999999999999999999999999999999999999999.1").hasInfiniteValue());
+        assertTrue(Num.of("9999999999999999999999999999999999999999999999999999999999999999999999999999999.0").hasInfiniteValue());
+    }
 }
